Change restart_count type to i32 (#179)
https://github.com/kubernetes/client-go/blob/f6ce18ae578c8cca64d14ab9687824d9e1305a67/applyconfigurations/core/v1/containerstatus.go#L28

+++ src/k8-types/src/core/pod.rs
@@ -226,7 +226,7 @@ pub struct ContainerStatus {
     pub name: String,
     pub state: ContainerState,
     pub ready: bool,
-    pub restart_count: u8,
+    pub restart_count: i32,
     pub image: String,
     #[serde(rename = "imageID")]
     pub image_id: String,
